Minimap working

+++ Minimap.hpp
@@ -11,17 +11,17 @@ class Minimap {
 
 	sf::Uint8 *pixels;
 public:
-	int width, height;
+	size_t width, height;
 
-	Minimap(int, int);
+	Minimap(size_t, size_t);
 	~Minimap();
-	
+
 	/* Sets minimap pixel at index i to color c */
 	void set(size_t, sf::Color color = sf::Color(22, 22, 22));
 	void set(size_t, size_t, sf::Color color = sf::Color(22, 22, 22));
 	void updateTexture();
 
-	void update(Grid &, int, int);
+	void update(Grid &, size_t, size_t);
 
 	void draw(sf::RenderTarget &);
 };
